@@ -57,63 +58,58 @@
     "popover"
   ],
   "peerDependencies": {
-    "react": "0.14.x || ^15.0.0 || ^16.0.0 || ^17.0.0"
+    "@popperjs/core": "^2.0.0",
+    "react": "^16.8.0 || ^17 || ^18",
+    "react-dom": "^16.8.0 || ^17 || ^18"
   },
   "dependencies": {
-    "@babel/runtime": "^7.1.2",
-    "@hypnosphi/create-react-context": "^0.3.1",
-    "deep-equal": "^1.1.1",
-    "popper.js": "^1.14.4",
-    "prop-types": "^15.6.1",
-    "typed-styles": "^0.0.7",
+    "react-fast-compare": "^3.0.1",
     "warning": "^4.0.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.2",
-    "@babel/core": "^7.1.2",
-    "@babel/plugin-external-helpers": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.1.0",
-    "@babel/plugin-transform-runtime": "^7.1.0",
-    "@babel/preset-env": "^7.1.0",
-    "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-stage-2": "^7.0.0",
-    "@types/react": "^16.1.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.0.0",
-    "cross-env": "^5.1.4",
-    "emotion": "^9.1.1",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
-    "enzyme-to-json": "^3.4.0",
-    "eslint": "^4.19.1",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-flowtype": "^2.46.1",
-    "eslint-plugin-jest": "^21.15.0",
-    "eslint-plugin-promise": "^3.7.0",
-    "eslint-plugin-react": "^7.7.0",
-    "flow-bin": "^0.69.0",
-    "flow-copy-source": "^1.3.0",
-    "gh-pages": "^1.1.0",
-    "git-branch-is": "^0.1.0",
-    "jest": "^24.0.0",
-    "parcel-bundler": "^1.10.1",
-    "prettier": "^1.11.1",
-    "pretty-quick": "^1.4.1",
-    "react": "16.9.0",
-    "react-dom": "^16.9.0",
-    "react-emotion": "^9.1.1",
+    "@atomico/rollup-plugin-sizes": "^1.1.3",
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.9.0",
+    "@babel/polyfill": "^7.8.7",
+    "@babel/preset-env": "^7.9.0",
+    "@babel/preset-flow": "^7.9.0",
+    "@babel/preset-react": "^7.9.4",
+    "@emotion/core": "^10.0.28",
+    "@emotion/styled": "^10.0.27",
+    "@popperjs/core": "^2.3.3",
+    "@rollup/plugin-commonjs": "^11.0.2",
+    "@rollup/plugin-node-resolve": "^7.1.1",
+    "@rollup/plugin-replace": "^2.3.1",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/react-hooks": "^8.0.0",
+    "@types/react": "^16.9.29",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^25.2.4",
+    "cross-env": "^7.0.2",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.1",
+    "eslint-plugin-flowtype": "^4.7.0",
+    "eslint-plugin-jest": "^23.8.2",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-react-hooks": "^3.0.0",
+    "flow-bin": "^0.176.2",
+    "flow-copy-source": "^2.0.9",
+    "gh-pages": "^2.2.0",
+    "git-branch-is": "^3.1.0",
+    "jest": "^25.2.4",
+    "parcel-bundler": "^1.12.4",
+    "prettier": "^2.0.2",
+    "pretty-quick": "^2.0.1",
+    "react": "18.0.0",
+    "react-dom": "^18.0.0",
     "react-spring": "^8.0.27",
-    "recompose": "^0.26.0",
-    "rimraf": "^2.6.2",
-    "rollup": "^0.60.0",
-    "rollup-plugin-babel": "^4.0.3",
-    "rollup-plugin-commonjs": "^9.1.3",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-size-snapshot": "^0.5.1",
-    "rollup-plugin-uglify": "^4.0.0",
-    "typescript": "^2.8.1"
+    "react-test-renderer": "^18.0.0",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.3.1",
+    "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-terser": "^5.3.0",
+    "typescript": "^3.8.3"
   }
 }
