@@ -1,8 +1,9 @@
 {
   "name": "react-transition-group",
-  "version": "2.9.0",
+  "version": "4.4.5",
   "description": "A react component toolset for managing animations",
-  "main": "index.js",
+  "main": "cjs/index.js",
+  "module": "esm/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/reactjs/react-transition-group.git"
@@ -27,19 +28,37 @@
     "setupFiles": [
       "./test/setup.js"
     ],
+    "setupFilesAfterEnv": [
+      "./test/setupAfterEnv.js"
+    ],
     "roots": [
       "<rootDir>/test"
     ]
   },
   "peerDependencies": {
-    "react": ">=15.0.0",
-    "react-dom": ">=15.0.0"
+    "react": ">=16.6.0",
+    "react-dom": ">=16.6.0"
   },
   "dependencies": {
-    "dom-helpers": "^3.4.0",
+    "@babel/runtime": "^7.5.5",
+    "dom-helpers": "^5.0.1",
     "loose-envify": "^1.4.0",
-    "prop-types": "^15.6.2",
-    "react-lifecycles-compat": "^3.0.4"
+    "prop-types": "^15.6.2"
+  },
+  "release": {
+    "pkgRoot": "lib",
+    "verifyConditions": [
+      "@semantic-release/changelog",
+      "semantic-release-alt-publish-dir",
+      "@semantic-release/git",
+      "@semantic-release/github"
+    ],
+    "prepare": [
+      "@semantic-release/changelog",
+      "semantic-release-alt-publish-dir",
+      "@semantic-release/npm",
+      "@semantic-release/git"
+    ]
   },
   "browserify": {
     "transform": [
