@@ -260,53 +261,57 @@
   "homepage": "https://github.com/reactstrap/reactstrap#readme",
   "dependencies": {
     "@babel/runtime": "^7.12.5",
+    "@popperjs/core": "^2.6.0",
     "classnames": "^2.2.3",
     "prop-types": "^15.5.8",
-    "react-popper": "^1.3.6",
-    "react-transition-group": "^2.3.1"
+    "react-popper": "^2.2.4",
+    "react-transition-group": "^4.4.2"
   },
   "peerDependencies": {
-    "react": ">=16.3.0",
-    "react-dom": ">=16.3.0"
+    "react": ">=16.8.0",
+    "react-dom": ">=16.8.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.12.10",
     "@babel/core": "^7.12.10",
+    "@babel/eslint-parser": "^7.18.2",
     "@babel/plugin-proposal-export-default-from": "^7.12.1",
     "@babel/plugin-proposal-export-namespace-from": "^7.12.1",
     "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
-    "@babel/plugin-transform-runtime": "^7.2.0",
-    "@babel/preset-env": "^7.12.11",
     "@babel/preset-react": "^7.12.10",
+    "@storybook/addon-actions": "^6.3.7",
+    "@storybook/addon-essentials": "^6.3.7",
+    "@storybook/addon-links": "^6.3.7",
+    "@storybook/addon-storysource": "^6.3.7",
+    "@storybook/react": "^6.3.7",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "12.1.5",
+    "@testing-library/user-event": "13.5.0",
     "@types/react": "^16.9.51",
     "babel-eslint": "^9.0.0",
     "babel-loader": "^8.2.2",
-    "bootstrap": "^4.4.1",
-    "clean-webpack-plugin": "^3.0.0",
+    "babel-preset-react-app": "^10.0.1",
+    "bootstrap": "^5.1.0",
     "conventional-changelog-cli": "^2.0.21",
-    "conventional-recommended-bump": "^0.3.0",
-    "copy-webpack-plugin": "^6.0.2",
-    "coveralls": "^3.1.0",
-    "cross-env": "^2.0.0",
-    "css-loader": "^2.1.0",
-    "customize-cra": "^0.2.8",
+    "conventional-recommended-bump": "^4.1.1",
+    "cross-env": "^4.0.0",
+    "customize-cra": "^0.9.1",
     "dtslint": "^4.0.4",
     "ejs": "^2.5.9",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "eslint": "^5.0.0",
-    "eslint-config-react-app": "^3.0.6",
-    "eslint-plugin-flowtype": "^3.13.0",
-    "eslint-plugin-import": "^2.0.0",
-    "eslint-plugin-jsx-a11y": "^6.0.0",
-    "eslint-plugin-react": "^7.0.0",
-    "extract-text-webpack-plugin": "^3.0.2",
+    "enzyme-adapter-react-16": "^1.15.6",
+    "eslint": "^8.17.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-storybook": "^0.6.1",
     "holderjs": "^2.9.3",
     "json-loader": "^0.5.7",
-    "mini-css-extract-plugin": "^0.9.0",
-    "ncp": "^2.0.0",
-    "raw-loader": "^1.0.0",
-    "react": "^16.3.2",
+    "microbundle": "^0.13.3",
+    "prettier": "2.7.1",
+    "react": "^16.8.0",
     "react-app-rewired": "^1.6.2",
     "react-dom": "^16.3.2",
     "react-helmet": "^5.0.3",
