@@ -1,10 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import 'bootstrap/dist/css/bootstrap.css';
 import './index.css';
 import App from './App';
 
-ReactDOM.render(
-    <App/>,
-    document.getElementById('app')
-);
+const container = document.getElementById('app');
+const root = createRoot(container)
+root.render(<App />);
+
+// ReactDOM.createRoot(
+//     <App/>,
+//     document.getElementById('app')
+// );
